requirements.txt
annotated-types==0.7.0
asn1crypto==1.5.1
Authlib==1.3.0
awswrangler==3.7.3
bandit==1.7.8
black==24.4.2
boto3==1.34.115
botocore==1.34.115
build==1.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.5.3
cryptography==42.0.7
distlib==0.3.8
dparse==0.6.4b0
filelock==3.14.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
moto==5.0.8
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pbr==6.0.0
pg8000==1.31.2
pip-tools==7.4.1
platformdirs==4.2.2
pluggy==1.5.0
pyarrow==16.1.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pyproject_hooks==1.1.0
pytest==8.2.1
pytest-cov==5.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-testdox==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
responses==0.25.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
safety==3.2.0
safety-schemas==0.0.2
scramp==1.4.5
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
stevedore==5.2.0
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
Werkzeug==3.0.3
wheel==0.43.0
xmltodict==0.13.0
